+++ src/lib/index.ts
@@ -3,8 +3,8 @@ import path from 'path';
 import { UguisuError } from './misc/errors.js';
 import { UguisuOptions } from './misc/options.js';
 import { getDefaultProjectInfo, parseProjectFile, ProjectInfo } from './project-file.js';
+import { RunningEnv } from './running/common.js';
 import { run } from './running/run.js';
-import { RunningEnv } from './running/tools.js';
 import { analyze } from './semantics/analyze.js';
 import { AnalysisEnv } from './semantics/common.js';
 import { parse } from './syntax/parse.js';
