+++ src/lib/running/value.ts
@@ -1,125 +1,28 @@
 import { UguisuError } from '../misc/errors.js';
 import { UguisuOptions } from '../misc/options.js';
-import { Trace } from '../misc/trace.js';
-import { ProjectInfo } from '../project-file.js';
 import { FunctionDecl } from '../syntax/node.js';
+import { RunningEnv, Symbol } from './common.js';
 
-export class RunContext {
-    env: RunningEnv;
-    options: UguisuOptions;
-    projectInfo: ProjectInfo;
+export type Value =
+    | NoneValue
+    | NumberValue
+    | BoolValue
+    | StringValue
+    | CharValue
+    | StructValue
+    | ArrayValue
+    | FunctionValue;
 
-    constructor(env: RunningEnv, options: UguisuOptions, projectInfo: ProjectInfo) {
-        this.env = env;
-        this.options = options;
-        this.projectInfo = projectInfo;
-    }
-}
-
-export class RunningEnv {
-    layers: Map<string, Symbol>[];
-    trace?: Trace;
-
-    constructor(baseEnv?: RunningEnv, trace?: Trace) {
-        this.trace = trace;
-        if (baseEnv != null) {
-            this.layers = [...baseEnv.layers];
-        } else {
-            this.layers = [new Map()];
-        }
-    }
-
-    declare(name: string, initialValue?: Value) {
-        this.trace?.log(`declare symbol: ${name} ${initialValue}`);
-        this.layers[0].set(name, new Symbol(initialValue));
-    }
-
-    lookup(name: string): Symbol | undefined {
-        this.trace?.log(`get symbol: ${name}`);
-        for (const layer of this.layers) {
-            const symbol = layer.get(name);
-            if (symbol != null) {
-                return symbol;
-            }
-        }
-        return undefined;
-    }
-
-    enter() {
-        this.trace?.log(`enter scope`);
-        this.layers.unshift(new Map());
-    }
-
-    leave() {
-        this.trace?.log(`leave scope`);
-        if (this.layers.length <= 1) {
-            throw new UguisuError('Left the root layer.');
-        }
-        this.layers.shift();
-    }
-}
-
-export class Symbol {
-    value?: Value;
-    constructor(value?: Value) {
-        this.value = value;
-    }
-}
-
-export type EvalResult<T> = OkResult<T> | ReturnResult | BreakResult;
-
-export type OkResult<T> = {
-    code: 'ok',
-    value: T,
-};
-
-export type ReturnResult = {
-    code: 'return',
-    value: Value,
-};
-
-export type BreakResult = {
-    code: 'break',
-};
-
-export function createOk<T>(value: T): EvalResult<T> {
-    return { code: 'ok', value };
-}
-
-export function createReturn<T>(value: Value): EvalResult<T> {
-    return { code: 'return', value };
-}
-
-export function createBreak<T>(): EvalResult<T> {
-    return { code: 'break' };
-}
-
-export function isOk<T>(x: EvalResult<T>): x is OkResult<T> {
-    return (x.code == 'ok');
-}
-
-export function isReturn<T>(x: EvalResult<T>): x is ReturnResult {
-    return (x.code == 'return');
-}
-
-export function isBreak<T>(x: EvalResult<T>): x is BreakResult {
-    return (x.code == 'break');
-}
-
-//#region Values
-
-export type Value = NoneValue | NumberValue | BoolValue | StringValue | CharValue | StructValue | ArrayValue | FunctionValue;
-
-export type ValueOf<T extends ValueKind> =
-    T extends 'NoneValue' ? NoneValue :
-    T extends 'NumberValue' ? NumberValue :
-    T extends 'BoolValue' ? BoolValue :
-    T extends 'StringValue' ? StringValue :
-    T extends 'CharValue' ? CharValue :
-    T extends 'StructValue' ? StructValue :
-    T extends 'ArrayValue' ? ArrayValue :
-    T extends 'FunctionValue' ? FunctionValue :
-    never;
+export type ValueOf<T extends ValueKind>
+    = T extends 'NoneValue' ? NoneValue
+    : T extends 'NumberValue' ? NumberValue
+    : T extends 'BoolValue' ? BoolValue
+    : T extends 'StringValue' ? StringValue
+    : T extends 'CharValue' ? CharValue
+    : T extends 'StructValue' ? StructValue
+    : T extends 'ArrayValue' ? ArrayValue
+    : T extends 'FunctionValue' ? FunctionValue
+    : never;
 
 export function getTypeName(valueKind: ValueKind): string {
     switch (valueKind) {
@@ -305,5 +208,3 @@ export type NativeFuncHandler = (args: Value[], options: UguisuOptions) => Value
 export function isFunctionValue(x: Value): x is FunctionValue {
     return (x instanceof FunctionValue);
 }
-
-//#endregion Values
